+++ NPS Browser/AppDelegate.swift
@@ -17,14 +17,11 @@ let log = SwiftyBeaver.self
 class AppDelegate: NSObject, NSApplicationDelegate, NSUserNotificationCenterDelegate {
     
     lazy var downloadManager: DownloadManager = DownloadManager()
-    lazy var bookmarkManager: BookmarkManager = BookmarkManager()
 
     func applicationDidFinishLaunching(_ aNotification: Notification) {
         // Insert code here to initialize your application
         setupSwiftyBeaverLogging()
         setupDownloadsDirectory()
-
-//        populateMasterViewTable()
     }
     
     func applicationWillTerminate(_ aNotification: Notification) {
@@ -39,11 +36,9 @@ class AppDelegate: NSObject, NSApplicationDelegate, NSUserNotificationCenterDele
     func setupSwiftyBeaverLogging() {
         let console = ConsoleDestination()
         let file = FileDestination()
-        let cloud = SBPlatformDestination(appID: "Rl12gv", appSecret: "bPi6vxoi2hV7mxnAtlriT3kBXbmofepc", encryptionKey: "qkiqm2IcjwgzuCcmu6itctXna2nxLztf")
 
         log.addDestination(console)
         log.addDestination(file)
-        log.addDestination(cloud)
     }
     
     func setupDownloadsDirectory() {
@@ -51,19 +46,7 @@ class AppDelegate: NSObject, NSApplicationDelegate, NSUserNotificationCenterDele
         let dlDirName = "NPS Downloads"
         
         try! Folder(path: dlFolder.path).createSubfolderIfNeeded(withName: dlDirName)
-        log.info("Downloads folder created at \(dlFolder.appendingPathComponent(dlDirName, isDirectory: true).path)")
     }
-    
-//    func populateMasterViewTable() {
-//        if (coreDataIO.recordsAreEmpty()) {
-//            NetworkManager().makeHTTPRequest()
-//        } else {
-//            let content = coreDataIO.getRecords()
-//            Helpers().getDataController().setArrayControllerContent(content: content)
-//        }
-        
-//        let items = 
-//    }
     
     // MARK: - Notifications