+++ NPS Browser/db/LayerProtocols.swift
@@ -1,0 +1,14 @@
+//
+//  LayerProtocols.swift
+//  NPS Browser
+//
+//  Created by JK3Y on 8/11/18.
+//  Copyright © 2018 JK3Y. All rights reserved.
+//
+
+import RealmSwift
+
+public protocol Storable {
+}
+extension Object: Storable {
+}